Move custom theme into its own folder.
