+++ src/theme/index.js
@@ -1,0 +1,12 @@
+import getMuiTheme from 'material-ui/styles/getMuiTheme';
+
+const customTheme = getMuiTheme({
+  'palette': {
+    'primary1Color': '#006064',
+    'primary2Color': '#1a237e',
+    'accent1Color': '#bf360c',
+    'pickerHeaderColor': '#00838f'
+  }
+});
+
+export default customTheme;
